shadow-dom/shadow-trees/upper-boundary-encapsulation/test-010.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_04_01_10</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Mikhail Fursov" href="mailto:mfursov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#upper-boundary-encapsulation">
<meta name="assert" content="Upper-boundary encapsulation: The parentNode and parentElement attributes of the shadow root object must always return null.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
//check parentNode of usual shadow
test(function () {
    var d = newHTMLDocument();
    var s = createSR(d.body);

    assert_equals(s.parentNode, null, 'the parentNode attribute of the shadow ' +
        'root object must always return null');

}, 'A_04_01_10_T01');

//check parentElement of usual shadow
test(function () {
    var d = newHTMLDocument();
    var s = createSR(d.body);

    assert_equals(s.parentElement, null, 'the parentElement attribute of the shadow root object ' +
        'must always return null');

}, 'A_04_01_10_T02');

//check parentNode for nested shadow
test(function () {
    var d = newHTMLDocument();
    var s1 = createSR(d.body);
    var e1 = d.createElement('div');
    s1.appendChild(e1);
    var s2 = createSR(e1);

    assert_equals(s2.parentNode, null, 'the parentNode attribute of the shadow ' +
        'root object must always return null');

}, 'A_04_01_10_T03');

//check parentElement for nested shadow
test(function () {
    var d = newHTMLDocument();
    var s1 = createSR(d.body);
    var e1 = d.createElement('div');
    s1.appendChild(e1);
    var s2 = createSR(e1);

    assert_equals(s2.parentElement, null, 'the parentElement attribute of the shadow root object ' +
        'must always return null');

}, 'A_04_01_10_T04');
</script>
</body>
</html>
